move server spec util into spec file
